refactor(rich-text): self-contained palette scrollbar styling + cleanup from review
