+++ packages/ui/rich-text/src/at-panel/mentionPalette.css
@@ -22,6 +22,31 @@
   transition: opacity 160ms ease-in-out;
 }
 
+.rich-text-at-mention-palette-scrollbar-thumb {
+  width: 4px;
+  min-height: 24px;
+  border-radius: 999px;
+  background: var(--transparency-block);
+  cursor: grab;
+  transition: background-color 160ms ease-in-out;
+  will-change: transform;
+}
+
+.rich-text-at-mention-palette-scrollbar:hover
+  .rich-text-at-mention-palette-scrollbar-thumb {
+  background: var(--transparency-hover);
+}
+
+.rich-text-at-mention-palette-scrollbar[data-dragging="true"] {
+  opacity: 1;
+}
+
+.rich-text-at-mention-palette-scrollbar[data-dragging="true"]
+  .rich-text-at-mention-palette-scrollbar-thumb {
+  background: var(--transparency-hover);
+  cursor: grabbing;
+}
+
 .rich-text-at-mention-palette {
   container-type: inline-size;
 }
