+++ packages/ui/rich-text/src/at-panel/types.ts
@@ -1,7 +1,5 @@
 import type {
   RichTextAtInsertResult,
-  RichTextAtProvider,
-  RichTextAtProviderContext,
   RichTextAtQueryMatch
 } from "../types/at.ts";
 
@@ -47,50 +45,4 @@ export interface RichTextAtSearchGroup<TItem = unknown> {
   visibleCount: number;
   hasMore: boolean;
   emptyLabel?: string;
-}
-
-export type RichTextAtSearchState<TItem = unknown> =
-  | {
-      status: "idle";
-      query: string;
-      mode: "browse";
-      filter: RichTextAtFilterId;
-      categories: readonly RichTextAtFilterTab[];
-      groups: readonly RichTextAtSearchGroup<TItem>[];
-      error: null;
-    }
-  | {
-      status: "loading" | "ready";
-      query: string;
-      mode: "results";
-      filter: RichTextAtFilterId;
-      categories: readonly RichTextAtFilterTab[];
-      groups: readonly RichTextAtSearchGroup<TItem>[];
-      error: null;
-    }
-  | {
-      status: "error";
-      query: string;
-      mode: "results";
-      filter: RichTextAtFilterId;
-      categories: readonly RichTextAtFilterTab[];
-      groups: readonly RichTextAtSearchGroup<TItem>[];
-      error: string;
-    };
-
-export interface RichTextAtSearchInput {
-  workspaceId?: string | null;
-  currentUserId?: string | null;
-  query: string;
-  context?: RichTextAtProviderContext;
-}
-
-export interface RichTextAtSearchControllerOptions {
-  richTextAtProviders?: readonly RichTextAtProvider[];
-  providerGroups?: readonly RichTextAtProviderGroup[];
-  filterTabs?: readonly RichTextAtFilterTab[];
-  defaultFilterId?: RichTextAtFilterId;
-  debounceMs?: number;
-  pageSize?: number;
-  target?: string;
 }
